#162: Move cog check to correct position and put class YTDLSource in its own file
